+++ backend/bank-service/src/main/java/com/c1736/bankservice/repository/IAccountBankRepository.java
@@ -1,6 +1,7 @@
 package com.c1736.bankservice.repository;
 
 import com.c1736.bankservice.entities.AccountBank;
+import com.c1736.bankservice.entities.AccountTransfer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
