--- backend/bank-service/src/main/java/com/c1736/bankservice/service/AccountBankService.java
@@ -1,11 +1,0 @@
-package com.c1736.bankservice.service;
-
-import com.c1736.bankservice.service.dto.request.TransferRequestDto;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface AccountBankService {
-
-    public void Transfer(TransferRequestDto transferRequestDto);
-
-}
