+++ backend/bank-service/src/main/java/com/c1736/bankservice/service/exceptions/AccountBankNotFound.java
@@ -1,7 +1,7 @@
 package com.c1736.bankservice.service.exceptions;
 
 public class AccountBankNotFound extends RuntimeException{
-    public AccountBankNotFound(String cuentaDeOrigenNoEncontrada) {
+    public AccountBankNotFound() {
         super();
     }
 }
