+++ backend/bank-service/src/main/java/com/c1736/bankservice/service/impl/AccountBankService.java
@@ -4,9 +4,12 @@ import com.c1736.bankservice.client.IMessagingFeignClient;
 import com.c1736.bankservice.client.dto.MessagingDTO;
 import com.c1736.bankservice.configuration.Constants;
 import com.c1736.bankservice.entities.AccountBank;
+import com.c1736.bankservice.entities.AccountTransfer;
+import com.c1736.bankservice.repository.AccountTransferRepository;
 import com.c1736.bankservice.repository.IAccountBankRepository;
 import com.c1736.bankservice.service.IAccountBankService;
 import com.c1736.bankservice.service.dto.request.AccountBankRequestDto;
+import com.c1736.bankservice.service.dto.request.TransferRequestDto;
 import com.c1736.bankservice.service.dto.request.UpdateAccountBankRequestDto;
 import com.c1736.bankservice.service.dto.response.AccountBankResponseDto;
 import com.c1736.bankservice.service.exceptions.AccountBankNotFound;
@@ -17,6 +20,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -28,17 +32,19 @@ public class AccountBankService implements IAccountBankService {
     private final IAccountBankRequestMapper accountBankRequestMapper;
     private final IAccountBankResponseMapper accountBankResponseMapper;
     private final IMessagingFeignClient messagingFeignClient;
+    private final AccountTransferRepository accountTransferRepository;
 
-    public AccountBankService(IAccountBankRepository accountBankRepository, IAccountBankRequestMapper accountBankRequestMapper, IAccountBankResponseMapper accountBankResponseMapper, IMessagingFeignClient messagingFeignClient) {
+    public AccountBankService(IAccountBankRepository accountBankRepository, IAccountBankRequestMapper accountBankRequestMapper, IAccountBankResponseMapper accountBankResponseMapper, IMessagingFeignClient messagingFeignClient, AccountTransferRepository accountTransferRepository) {
         this.accountBankRepository = accountBankRepository;
         this.accountBankRequestMapper = accountBankRequestMapper;
         this.accountBankResponseMapper = accountBankResponseMapper;
         this.messagingFeignClient = messagingFeignClient;
+        this.accountTransferRepository = accountTransferRepository;
     }
 
     @Override
     public AccountBankResponseDto getAccountBank(Long id) {
-        AccountBank account = accountBankRepository.findById(id).orElseThrow(() -> new AccountBankNotFound("Cuenta de origen no encontrada"));
+        AccountBank account = accountBankRepository.findById(id).orElseThrow(() -> new AccountBankNotFound());
         return accountBankResponseMapper.toResponseAccountBank(account);
     }
 
@@ -46,7 +52,7 @@ public class AccountBankService implements IAccountBankService {
     public List<AccountBankResponseDto> getAllAccountBank() {
         List<AccountBank> accountBanks = accountBankRepository.findAll();
         if (accountBanks.isEmpty()) {
-            throw new AccountBankNotFound("Cuenta de origen no encontrada");
+            throw new AccountBankNotFound();
         }
         return accountBanks.stream()
                 .map(accountBankResponseMapper::toResponseAccountBank)
@@ -82,7 +88,7 @@ public class AccountBankService implements IAccountBankService {
             accountBank.setTypeCoin(updateAccountBankRequestDto.getTypeCoin());
             accountBankRepository.save(accountBank);
         } else {
-            throw new AccountBankNotFound("Cuenta de origen no encontrada");
+            throw new AccountBankNotFound();
         }
     }
 
@@ -91,7 +97,36 @@ public class AccountBankService implements IAccountBankService {
         try {
             accountBankRepository.deleteById(id);
         } catch (EmptyResultDataAccessException e) {
-            throw new AccountBankNotFound("Cuenta de origen no encontrada");
+            throw new AccountBankNotFound();
         }
     }
+
+    @Override
+    public void Transfer(TransferRequestDto transferRequestDto) {
+        AccountBank fromAccount =accountBankRepository.findById(transferRequestDto.getFromAccount())
+                .orElseThrow(AccountBankNotFound::new);
+
+        AccountBank toAccount = accountBankRepository.findById(transferRequestDto.getToAccount())
+                .orElseThrow(AccountBankNotFound::new);
+
+        if(fromAccount.getBalance().compareTo(transferRequestDto.getAmount())<0){
+            throw new IllegalArgumentException("Saldo insuficiente para realizar la transferencia");
+        }
+
+        fromAccount.setBalance(fromAccount.getBalance().subtract(transferRequestDto.getAmount()));
+        toAccount.setBalance(toAccount.getBalance().add(transferRequestDto.getAmount()));
+
+        AccountTransfer transfer = new AccountTransfer();
+        transfer.setFromAccount(fromAccount);
+        transfer.setToAccount(toAccount);
+        transfer.setAmount(transferRequestDto.getAmount());
+        transfer.setDescription(transferRequestDto.getDescription());
+        transfer.setTransferDate(LocalDateTime.now());
+
+        accountTransferRepository.save(transfer);
+        accountBankRepository.save(fromAccount);
+        accountBankRepository.save(toAccount);
+
+    }
+
 }
